Revert "Added funktion to all actionbars, Added EnergySystem"
This reverts commit b74710057963ce4cb2803cb76f19d81a1317ab2b.

+++ spel_modul2/GameEngine/Components/CoreComponents/EnergyComponent.cs
@@ -2,10 +2,10 @@
 {
     public class EnergyComponent : IComponent
     {
-        public float Max { get; set; }
-        public float Current { get; set; }
+        public int Max { get; set; }
+        public int Current { get; set; }
 
-        public EnergyComponent(float max)
+        public EnergyComponent(int max)
         {
             Max = max;
             Current = Max;
